+++ test/contracts/LegacyUUPSV2.sol
@@ -1,0 +1,13 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.22;
+
+import {LegacyUUPS} from "./LegacyUUPS.sol";
+
+contract LegacyUUPSV2 is LegacyUUPS {
+    uint256 public revision;
+
+    function increment() external {
+        ++value;
+        revision = 2;
+    }
+}